requirements.txt


pandas>=1.0.0,<2.0.0
psycopg2-binary==2.8.6
python-dotenv==0.19.2
SQLAlchemy==1.3.24
python-dotenv